pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ortholoc"
version = "1.0.0"
description = "OrthoLoC Python library"
authors = [
  { name = "DeepScenario 2025 - Oussema Dhaouadi" }
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "CC BY-NC-SA 4.0"}
dependencies = [
  "roma",
  "loguru",
  "imcui",
  "torch",
  "rasterio",
  "py-cpuinfo",
  "gputil",
  "psutil",
  "opencv-python-headless",
  "torchvision",
  "tueplots",
  "packaging",
  "beautifulsoup4",
  "appdirs"
]

[project.optional-dependencies]
gpu = [
  "torch==2.0.0+cu118",
  "torchvision==0.15.0+cu118"
]

[project.scripts]
run-benchmark = "ortholoc.scripts.run_benchmark:main"
run-localization = "ortholoc.scripts.run_localization:main"
run-matching = "ortholoc.scripts.run_matching:main"
visualize-sample = "ortholoc.scripts.visualize_sample:main"
visualize-dataset = "ortholoc.scripts.visualize_dataset:main"

[tool.hatch.build]
include = [
  "ortholoc/*"
]

[project.urls]
Homepage = "https://deepscenario.github.io/OrthoLoC"
Bug_Tracker = "https://github.com/deepscanrio/OrthoLoC/issues"
